third_party/blink/web_tests/virtual/plz-service-worker/http/tests/serviceworker/clients-openwindow-expected.txt

This is a testharness.js-based test.
FAIL clients.openWindow() behaved as expected assert_array_equals: Worker should post back expected messages. expected property 10 to be " focused: false" but got " focused: true" (expected array ["openWindow() can't open a window without a user interaction", "openWindow() error is InvalidAccessError", "openWindow() can open cross origin windows", "openWindow() result: null", "openWindow() can open not controlled windows", "openWindow() result: [object WindowClient]", "openWindow() can open controlled windows", "openWindow() result: [object WindowClient]", " url: http://127.0.0.1:8000/serviceworker/resources/blank.html", " visibilityState: visible", " focused: false", " frameType: top-level", "openWindow() can open about:blank", "openWindow() result: null", "openWindow() can open about:crash", "openWindow() result: null", "openWindow() can not open an invalid url", "openWindow() error is: TypeError", "openWindow() can not open view-source scheme", "openWindow() error is: TypeError", …] got ["openWindow() can't open a window without a user interaction", "openWindow() error is InvalidAccessError", "openWindow() can open cross origin windows", "openWindow() result: null", "openWindow() can open not controlled windows", "openWindow() result: [object WindowClient]", "openWindow() can open controlled windows", "openWindow() result: [object WindowClient]", " url: http://127.0.0.1:8000/serviceworker/resources/blank.html", " visibilityState: visible", " focused: true", " frameType: top-level", "openWindow() can open about:blank", "openWindow() result: null", "openWindow() can open about:crash", "openWindow() result: null", "openWindow() can not open an invalid url", "openWindow() error is: TypeError", "openWindow() can not open view-source scheme", "openWindow() error is: TypeError", …])
Harness: the test ran to completion.
